hw/arm/virt: Provide DT binding generation for PCI eXpander Bridges
This is a precursor series for CXL on arm virt (for which DT support was
requested) as CXL QEMU emulation uses pxb-cxl which inherits from the
slightly simpler pxb-pcie. ACPI support for these additional host bridges
has been available for some time but relies an interesting dance with
EDK2:
* During initial board creation etc the PXB buses are largely ignored
  and ACPI tables + DT passed to EDK2 only expose the section of the
  PXB instance that exists as a normal PCI EP on the main root bus.
* EDK2 then carries out full PCI bus enumeration, including the buses
  below the PXB host bridges.
* Finally EDK2 calls back into QEMU to rebuild the tables via
  virt_acpi_build_update, at which point the correct DSDT for the
  PXB host bridges is generate and an adjust DSDT section is generated
  for the primary host bridge.

For device tree bindings there is no such dance with the firmware
and as such we need QEMU to directly present device tree entries
for the primary PCIe bus and the PXB instances.

This cannot be fully done either at initial PCIe instantiation,
(PXB instances not yet instantiated) or at in virt_machine_done
as it is for ACPI (virtio-mem hot plug routines rely on presence
of primary PCIe bus).  Thus proposed solution is to set it up
initially without being careful and later update it with additional
checks that we don't get overlapping bus numbers.
It 'might' be possible to use hotplug handlers to update the
DT as we go along, but that would be rather challenging as
each additional PXB introduction would sometims require an
update of the dt for all other instances.

The linux,pci-domain dt element is more restrictive than the
equivalent ACPI handling, so for now just put each PXB in it's
own domain.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>