misc/i2c_mctp_cxl_fmapi: Initial device emulation
The Fabric Manager API is used to configure CXL switches and
devices. DMTF has defined an MCTP binding specification to carry
these messages. Then end goal of this work is to hook this
up to an emulated CXL switch and allow control of the switch
configuration.  For now it's stand alone to show how the MCTP side
of things works before hooking up the complexities of real switch
emulation.

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>