hw/i386/acpi: Separate PXB related parts of DSDT into an SSDT table.
The dependencies between the CRS entries and bus numbers due to
PCI eXpander Bridges taking resources from the primary bus make this
a slightly complex dance.

The rules we have to fit into are:
1) FACP is first table in RSDT and points to the DSDT.
2) Thus DSDT is created before FACP
3) SSDT must be pointed to by a later entry in RSDT so must
   be created after DSDT.
4) The main PCI bus in DSDT contains entries that are dependent on
   bus walks done to establish the entries for the PXB buses which
   are now in SSDT.

Solution is to precompute what will go in SSDT (CRS ranges and bus
numbers) then to create the DSDT + FACTP as normal.

After that create the SSDT, including rerunning some of the earlier
bus walking code (which will give the same answers).

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>