hw/cxl/i2c_mctp_cxl: Initial device emulation
The CCI and Fabric Manager APIs are used to configure CXL switches and
devices. DMTF has defined an MCTP binding specification to carry these
messages. The end goal of this work is to hook this up to emulated CXL
switches and devices to  allow control of the configuration.

Since this relies on i2c target mode, this can currently only be used with
an SoC that includes the Aspeed I2C controller.

Note, no commands supported at this point.

Signed-off-by: Jonathan Cameron <jonathan.cameron@huawei.com>