hw/cxl: Add scan media cmd support
Implement the poison list backup functionality by extending the
qmp poison injection interface to add chunks beyond the poison
limit and use it to recreate the poison list upon overflow.

To this end two new lists are added, bkp and results, which are
drained by moving elements from the backup to the results, which
are consumed by the respective get scan media results cmd.

Signed-off-by: Davidlohr Bueso <dave@stgolabs.net>
Link: https://lore.kernel.org/r/20230908073152.4386-4-dave@stgolabs.net